Update patch set 2
Patch Set 2:

I pulled in genius and most of the tests use the AbstractTestableListener which awaits async. A couple tests fail in InterfaceManagerConfigurationTest but they're not related to the AbstractConcurrentDataBrokerTest change as the failed assertions happen before verification of the AsyncEventsWaiter plus they fail even if I revert this change. The other failures were in MdSalUtilTest and those were related to these changes and I've pushed a patch to fix them. I'll look at netvirt next.

Patch-set: 2