Update patch set 2
Patch Set 2:

It doesn't look like netvirt can be easily changed to work with async DTCLs. The problem is with the AbstractTestableListener - it's intended to handle async change events but the mechanism of counting events via incrementing and decrementing and expecting the net count to be 0 is basically flawed when events are actually delivered async. For one, if awaitEventsConsumption is called before the first event is delivered, it simply returns - it has no idea that the expected event(s) simply haven't occurred yet. I added a CountDownLatch to detect the first event and it fixed a couple tests but many others still fail intermittently for other reasons. I think we need a simpler test DTCL that takes an expected event count and uses a CountDownLatch. But I'm not familiar with the netvirt code nor the various test frameworks in play (guice, the xtend stuff...) and I don't really have the time to invest in it so I think I'll have to make the MT DTCL executor an opt-in. So unfortunately the tests where deadlocks have occurred (as in bgpcep) will have to configure the MT executor.

Patch-set: 2
Label: Verified=0