Bug 8494: Reduce DEAD_TIMEOUT_SECONDS to 3m
This does not prevent Bug 8494 symptom,
but this helps investigating when the symptom occures.

AbstractTransactionHandler defines the subject constant,
which tells it when to give up waiting for future completion.

Currently, Bug 8494 is the only bug (occasionally) present
which causes produce-transactions to not return;
but its test case has short time window to fail,
so waiting full 15 minutes would make the test take too long
without much added benefit.

The new value of three minutes is safely larger than the usual
request timeout (2m), and it is within the waiting period
of current test shutdown.

Change-Id: If398a6d421b0f55d7a58dd101c0ce89232549801
Signed-off-by: Vratko Polak <vrpolak@cisco.com>