WIP: BUG-8733: speed up DataTree operations
Since the DataTree instance is used to merge data from this
and foreign shards, which do not have consistency guarantees, it
does not make sense to perform deep validation.

Migrate to new DataTreeFactory method and explicitly disable optional
validations. Also outlines further steps that need to be taken.

This flushes out a problem in test or in DataTree, where we fail
to find the root path in the SchemaContext. This needs to be checked.

Change-Id: I0fe0e7fdec5d4afb31c6ac493149542f5dfaa1bf
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>