WIP: guard against recursion in persistData()
This is a prototype patch ensuring we do not invoke the on-persist-complete
callback before we issued the replicate request.

NOTE: This needs to be checked for accounting correctness, because the recursive
case can make AbstractLeader.possiblyUpdateCommitIndex() see replication as
complete, but CheckConsensus has not been invoked, nor the snapshot capture
was not done.

Change-Id: Idec77f8af8eb6cf2d030b0c49b14e979491d11d4
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>