Bug 9060: TracingBroker printOpenTransactions
This is method is intended to be surfaced e.g. from a future Karaf CLI
command, and maybe JMX or something else like that which can be invoked
during future automated testing.

This is one of a serious of commits which is part of a solution I'm
proposing in order to be able to detect OOM issues such as Bug 9034,
based on using the mdsal-trace DataBroker.

Change-Id: I682700bef9644834e8b4ca36b21729f021a76bf0
Signed-off-by: Michael Vorburger <vorburger@redhat.com>