Update patch set 3
Patch Set 3: Code-Review-1

As far as I understand (Tom please correct me if I'm wrong, but this based on your clarification on the https://lists.opendaylight.org/pipermail/mdsal-dev/2017-August/001323.html post...), the currently proposed inline documentation is wrong: 

This property is NOT "useful for tracking down lifecycle offenders" (if by that we mean tracking non-closed transactions) - this flag does NOT help with that (but I'm building a new solution for that in Bug 9060).  What it does does do, AFAIK, is print the stack trace of the creator of the transaction in case there is an exception when the transaction is submitted (like a failed validation, perhaps?).

Maybe we could use inline doc saying something to that effect, and remove "lifecycle offenders" (perhaps even point to https://bugs.opendaylight.org/show_bug.cgi?id=9060 in the inline doc here?).

Patch-set: 3
Label: Code-Review=-1