Update patch set 11
Patch Set 11:

Not and archetype expert, but i was looking in the patch to see if we added support for blueprint maven plugin for annotation based DI to it, but i think i didn't find anything related. If it's there, please ignore the comment :), but if not, i think adding that will bring really great value.

Patch-set: 11
Label: Verified=0