Update patch set 11
Patch Set 11:

(2 comments)

> -1 just to clarify two points, which are alluded to in the commit
 > message's TBDs - but the commit message should explain why they're
 > disabled too (i.e., what's NOK about localFeature, and what's 
 > wrong with the it module).

Commit message updated.  I don't recall the exact failure message (and don't see much value in re-running just to provide that here now) - anyone wanting to work on this in the future will just uncomment it, and see. (Of course, as of right now, it will would because the "self test" is disabled, but I'm hoping to address that ASAP, like next week hopefully, and much before having any time and motivation to also look at the IT and the karaf.localFeature.  Unless they get fixed by the same whatever trick - we'll see.

Patch-set: 11
Label: Verified=0