Update patch set 1
Patch Set 1:

This is nice! You wouldn't want to cover this with TracingBrokerTest, would you? Just an idea, not -1 - I'll also merge it without that, but if you like TDD.. ;)

PS: The other places I mentioned which print out StackTraceElement where theoretically something like this could be cool too are Josh's getStackSummary() method in the same TracingBroker class, and if you grep for isTransactionDebugContextEnabled in Tom's sal-distributed-datastore (although AFAIK that does really dump/print itself, just include a Throwable as an Exception cause?).

Patch-set: 1
Label: Code-Review=0