Update patch set 1
Patch Set 1:

> > what do you want to do with it?
 > 
 > Push it forward.
 > By the way, what is the preferred way for me (as Controller
 > contributor) to let you (as Controller committer) know when I
 > consider a particular patch set to be ready for review and merge?

I'm not aware of any preferred/standard way but I'd say by adding a committer as a reviewer. If it's not actually ready to be merged then add "WIP" or "DO NOT MERGE" in the summary. Or push a draft patch until it's ready to be merged which is what I do.

Patch-set: 1
Reviewer: Gerrit User 1842 <1842@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Verified=0