Update patch set 15
Patch Set 15:

(1 comment)

Robert FYI I consider that "final" which this adds everywhere as visual clutter, and I hope this is just your habit and not something one of our Checkstyle changes in odlparent 3.0.0 mandates, or is it?  With error-prone and its @Var I'm hoping to eventually move us away from "final" on all args and local vars - already done so in infrautils with error-prone.  That's not a reason to -1 this though, so I won't - more FYI.

Patch-set: 15