CONTROLLER-1764: WIP: avoid refreshes
odl-mdsal-trace can no longer be installed with Karaf 4, because of a
number of bundle refreshes which result in the datastore restarting
while changing classloaders, resulting in chaos. This patch enforces a
number of optional dependencies, avoiding the refreshes.

Note that it currently pulls in odlparent 2.0.6-SNAPSHOT for
odl-mdsal-clustering-commons, to avoid ODLPARENT-132.

Change-Id: I8f5c75c454c40a19f9a4996571f6f8505de00fb0
Signed-off-by: Stephen Kitt <skitt@redhat.com>