Update patch set 1
Patch Set 1: Code-Review-1

This is technically an API change and we're past the freeze date for that so I think we should wait on this.  Also every downstream project that uses FindBugs should be built with this patch and any violations fixed before this patch gets merged.

Patch-set: 1
Reviewer: Gerrit User 1842 <1842@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Code-Review=-1