Update patch set 1
Patch Set 1:

> This is technically an API change and we're past the freeze date
 > for that so I think we should wait on this.  

Can we discuss, clarify and find binding agreement what we should wait for?  Are we e.g. saying that there is no way that we can still do this in Oxygen anymore, but we're cool to make this change at the very beginning of the opening of the new master branch post Oxygen?  Let's just spell out exactly what we're stating here... ;-)

 > Also every downstream
 > project that uses FindBugs should be built with this patch and any
 > violations fixed before this patch gets merged.

I could and will volunteer to do that, but let us first find an agreement on above?

Patch-set: 1
Label: Verified=0