Update patch set 1
Patch Set 1:

> > This is technically an API change and we're past the freeze date
 > > for that so I think we should wait on this.
 > 
 > Can we discuss, clarify and find binding agreement what we should
 > wait for?  Are we e.g. saying that there is no way that we can
 > still do this in Oxygen anymore, but we're cool to make this change
 > at the very beginning of the opening of the new master branch post
 > Oxygen?  Let's just spell out exactly what we're stating here...
 > ;-)
 > 
 
yes that is it. I believe we would technically have to go thru the API waiver protocol to get it into Oxygen now.  We can discuss doing this in Oxygen on the kernel call. I'd be OK getting it in  but I don't want to get yelled at as PTL should it break someone downstream :)

Patch-set: 1