Update patch set 1
Patch Set 1:

> Talking with Tom, also seems better to just change the default to
 > disabled. There really aren't many projects (if any) in ODL that
 > can truly restart. If a project needs the functionality, they can
 > flip the flags in their own controlled cfg files.

The default is disabled, ie restarting is an opt-in. I think permanently disabling these in the BP xml  is good in general as you said, especially the broker-impl one as that will restart the world.  I think we should also comment out the code for now since it's broken.

Patch-set: 1