Update patch set 4
Patch Set 4: Code-Review+2

> > Maybe should just remove the it code? Noone beyond experts really
 > > use it properly so it is distracting to new users. And it doesn't
 > > really test anything beyond sft or simply loading your feature.
 > 
 > I agree it's noise, along with CLI (https://git.opendaylight.org/gerrit/#/c/67070/)
 > et al.  The startup archetype should be just that  - get someone
 > started. Let's discuss this on the kernel call tomorrow.

As this finally "works" (OK with it/ commented out..), let me finally get this in as is.  We can always follow-up with removing more.  My personal view is that the CLI example bundle definitely has value - not sure why we should loose that.  I'm OK to ditch the IT, although it has value, I think we've found 2 real problems here with CONTROLLER-1810 and CONTROLLER-1811, but if everone is for ditching it.. fine with me. What I would LOVE to do is contribute an example of a component test to the archetype... (incl the required dependency to infrautils:inject.guice.testutils).

Patch-set: 4
Label: Code-Review=+2