Update patch set 2
Patch Set 2:

> > Yup, I explained that to Michael on IRC ;-).
 > 
 > and "someone" (I think it might have ben been me!) documented this
 > on https://wiki.opendaylight.org/view/BestPractices/Coding_Guidelines#IllegalCatch
 > ... ;-)  Tom FYI Stephen just made a useful addendum to that doc,
 > about considering existing throws clauses.

My first choice is to catch the appropriate ex types and not suppress. But sometimes you run into a method that declares throws Exception so you have no choice but to suppress.  In case where no checked ex's are actually thrown, I'm hesitant to remove the try/catch b/c I don't know the original intention. So much of mdsal and yangtools code is littered with paths that throw ISE and other unchecked even if the method declares a checked ex so, unfortunately, you pretty much have to catch RuntimeException everywhere to be safe.

Patch-set: 2