Update patch set 1
Patch Set 1:

> can we merge this change, since the initial code drop to archetypes
 > is merged?

I'm hoping to find time over the next week or two to raise a counter proposal to this one, where to reduce future confusion to end-users we keep the archetype itself, but empty, and direct to the new one.  Before doing that, I need to change the package in the new one though, otherwise it is even more confusing.

Patch-set: 1