Use equinox in AbstractMdsalTestBase
We're seeing an uptick in

  ServiceLookupException: gave up waiting for service org.ops4j.pax.exam.ProbeInvoker

failures in ITs. The surefire output file just shows this error:

  Error in initialization script: .../etc/shell.init.script: String index out of range: 0

which isn't much to go on but I also see that error in successful runs. We don't
need the karaf shell so we can disable it.

I did notice that it's using felix - we should be using equinox - let's see
if that alleviates the failures.

Change-Id: I73ded3c1d6f9a7b81ea25aa7de93b615e4e6ff26
Signed-off-by: Tom Pantelis <tompantelis@gmail.com>