Disable karaf console in AbstractMdsalTestBase
We're seeing an uptick in
ServiceLookupException: gave up waiting for service org.ops4j.pax.exam.ProbeInvoker
failures in ITs. The common denominator seems to be this error:

  Error in initialization script: .../etc/shell.init.script: String index out of range: 0

in the surefire output file. This isn't much to go on as to what the issue is
in the shell.init.script but we don't need the karaf shell so disable it. Hopefully
this alleviates the failures.

Change-Id: I73ded3c1d6f9a7b81ea25aa7de93b615e4e6ff26
Signed-off-by: Tom Pantelis <tompantelis@gmail.com>