make validate/prepare/commit methods in InMemoryDOMDataStore protected
instead of package private; because I'd like to @Override them in
something I am experimenting with (a POC where I would like to extend
InMemoryDOMDataStore to send the DataTreeCandidate to an external data
store on commit).

and make dataTree field protected, and add a
protected void notifyListeners(DataTreeCandidate candidate)
method.

Change-Id: I70a01172533c449ba70c8c2eb74315029cb4d6ae
Signed-off-by: Michael Vorburger <vorburger@redhat.com>