fix bug in NormalizedNodeInputStreamReader re. SIGNATURE_MARKER
and add testPathArgumentStreaming() for non-regression and to illustrate
it - it fails without that readSignatureMarkerAndVersionIfNeeded() added
in NormalizedNodeInputStreamReader.

The testPathArgumentStreaming() is based on the existing
testYangInstanceIdentifierStreaming(), but uses a PathArgument instead
of a YangInstanceIdentifier.

This probably doesn't fix anything actually broken in CDS, because it,
presumably (not verified) always uses another operation of
NormalizedNodeDataOutput (like writeSchemaPath, writeNormalizedNode or
writeYangInstanceIdentifier) before it does a writePathArgument, and so
the SIGNATURE_MARKER was already written.

I've only hit it in something where I directly wrote a PathArgument.

Change-Id: I30678b819adb8d485de8365e69939d983dc912de
Signed-off-by: Michael Vorburger <vorburger@redhat.com>