Proxy binding adapters to mdsal APIs
First, I apologize for the size of this patch. I tried to
break it up but was problematic and would've required some
throw-away transition code.

Some of the adapter classes were removed, others were renamed
from "BindingDOM*" to "Binding*" as they no longer adapt DOM APIs.
I did Refactor -> Move in eclipse but for some reason some appear
as renamed as desired but others appear as an add and delete.

mdsal-trace-dom was changed to implement the mdsal DOMDataBroker API
and also utilizes the LegacyDOMDataBrokerAdapter to provide the
controller API services. The mdsal-binding-dom-adapter bundle
automatically publishes the mdsal API binding services so
mdsal-trace-binding only needs to provide the controller API services.

Change-Id: If46e6ecb74a13f19be37661d0b55fabefbdf2edf
Signed-off-by: Tom Pantelis <tompantelis@gmail.com>