Update patch set 1
Abandoned

we were suspecting that we may loose this very important log message for some reason, but we will just apply something like this downstream to help debugging; there is actually no need for this to go into upstream master, so Abandon.  Sorry for the noise!

Patch-set: 1
Status: abandoned
Tag: autogenerated:gerrit:abandon