WIP: Purge ask-based transaction chains lazily
Since ask-based protocol does not have an explicit transaction chain
finalization message fired after all transactions are done, we need
to somehow schedule a purge.

We cannot schedule it immediately when a transaction chain is closed,
as there may still be outstanding transactions waiting to be
committed. We therefore create an alternative payload to
PurgeLocalHistoryPayload, which causes metadata builder to take over
and purge when it has no further transactions.

Change-Id: I1ec87a9905aa76f3e2a57ae2440c8c693ff60848
JIRA: CONTROLLER-1628
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>